+++ client/pages/Profile/Profile.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Register } from '../../components/Registration/Register'
+import Register from '../../components/Registration/Register'
 import { View } from '../../components/Registration/View'
 import { getState } from '../../store'
