--- server/.env.example
@@ -1,1 +1,0 @@
-SENDGRID_API_KEY='replace-with-api-key-from-sendgrid'
